+++ changelog.md
@@ -9,6 +9,9 @@ Imminent todos:
 - Add googley eyes to the animated metaballs demo.
 - check automatic darkmode detection
 
+* 2025-03-23
+  - Making the second parameter `center` of the `Vertex.rotate` method optional.
+  - Added the `Bounds.getMinDimension` and `Bounds.getMaxDimension` methods.
 * 2025-03-17 [v1.23.3]
   - Finalizing infinite irregular pattern demo.
 * 2025-02-12